Remove the need to calculate x_min_value in topK
The role of `x_min_value` was to prevent nodes not present in a graph to be selected in topK. This was done by setting the values in the positions outside the nodes in a graph as `x_min_value`-1. The PR applies `tanh` to score before passing to `topk` which bounds the scores between -1 and 1. Now by replacing `x_min_value` with -2 we achieve the same intended result.