Fix out of index issue in topk.
In topk, it makes dense and fixed node tensor according to x.shape and initialize all values to -2.
So, a graph which is smaller than biggest graph in batch has dummy nodes.
And then, x is copied to dense tensor and ordered by score.
If some scores in a graph which is smaller than biggest graph are smaller than -2,
they are moved to out of node size of original graph and dummy nodes are filled to those position.
So, returned perm has dummy node that have out of index.