Create learn_physic_model.py
Based on the arXiv paper: Relational inductive biases, deep learning, and graph networks.

Tutorial: This code proposes to deal with the problem of learning a physical model using
graphs and the torch geometric library. This work is based on the work done by
the GraphNets authors. The idea is to see how to use a pytorch based library
to achieve the same result as with the GraphNets library. Some simplifications
have been made compared to the approach proposed in the code using GraphNets.