Fix NeighborLoader bug with PyTorch-Lightning
When using PyTorch-Lightning's DDP on heterogeneous graphs, the NeighborLoader
implementation fails due to re-creation of the loader object. While Lightning
replaces the sampler to a distributed version, the `num_neighbors` parameter
is extracted from the current object and passed to the new one. At this point
it is already in the desired format, so we must not convert the dict keys again
using `edge_type_to_str()` function. This bug fix checks whether the `edge_type`
is already a string and does not apply the conversion in such case.