updated the __call__ of the NeighborSampler (#4098)
The call to the NeighborSampler in the torch_geometric/loader/neighbor_loader.py (line #63 on the master branch) has a bug at the beginning. The input argument "indices" is set to the new variable "index" if the "indices" is not a list. Then the assertion is triggered if the type does not match. The bug is for the case the input argument is indeed a torch.Tensor, hence the new variable "index" is undefined and the assert triggers an "Unknown Variable" error.

```python
def __call__(self, indices: List[int]):
    if not isinstance(indices, Tensor):
        index = torch.tensor(indices)
    assert index.dtype == torch.int64
```

The fix is simply just to change the the variable name "indices" to "index" at the beginning. Furthermore, the assertion can be also addressed properly as shown below:

```python
def __call__(self, index: Union[List[int], Tensor]):
    if not isinstance(index, torch.LongTensor):
        index = torch.LongTensor(index)
```