refactor(sampler): consolidate link neighbor sampling interface, part 2 (#5365)
This PR begins the effort to consolidate PyG's sampling interface in preparation for moving `sample(...)` behind the `GraphStore` interface. This effort is somewhat large in scope and will be broken into multiple PRs for ease of review. It resolves some TODOs from #5312, and introduces others that will be resolved in follow-up PRs. 

The major change removes `LinkNeighborSampler`, consolidating it under a the common `BaseSampler` interface with `NeighborSampler`. In doing so, it modifies the `BaseSampler` interface to support `sample_from_nodes` and `sample_from_edges`, since oftentimes edge-based sampling applies some transformations and then re-uses the same logic as node-based sampling. A sampler need not define both node-based and edge-based sampling, but if it doesn't, an appropriate error will be raised.

Resolved TODOs:
* `LinkNeighborSampler` and `NeighborSampler` interfaces are aligned
* No more special handling of `edge_type_to_str` in `filter_*`
* No more special handling of `perm_dict`, which removes the need for `edge_type_to_str` entirely