Update load_csv.py
load_node_csv() used  mapping = {index: i for i, index in enumerate(df.index.unique())} . But this will generate a mapping of the index of the dataframe to itself, since the index of the dataframe is a RangeIndex. To get the value from the index_col (the movie ID), enumerate(df[index_col]) should be used instead.