Add PyTorch SparseTensor support for `MessagePassing` (#5944)
This PR adds the PyTorch SparseTensor support for the base layer
`MessagePassing`. There are some points to be confirmed (as marked with
TODO):
+ ~~In `__collect__`: Since `adj._values()` returns a detached tensor,
should we use a coalesced matrix instead (e.g.,
`adj.coalesce().values()`)? This is for the case of computing sparse
gradients of `adj`.~~ (Solved)
+ In `__collect__`: Should we store the `ptr` for PyTorch SparseTensor
when fused aggregation is not available?
+ In `__lift__`: Should we use `gather_csr` for PyTorch SparseTensor?

Also, `torch.jit.script` is not available for PyTorch SparseTensor. Will
figure it out soon.

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>