Added the `GraphMaskExplainer` algorithm (#6284)
This PR contains implementation of how to compute `layer-wise weights`
for each `edge` in order to produce `explanations` for `node-level`,
`edge-level`, and `graph-level` tasks. Furthermore, this implementation
is different from authors' original implementation and is fast and more
memory efficient than theirs. Have added `Tests` and `Examples` of the
proposed implementation in order to make the overall approach
understandable.

Co-authored-by: fork123aniket <fork123aniket>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>