Add `GCN` and `FiLM` jit example (#6692)
This PR wants to make `example/film.py` and `example/gcn.py` jittable.
In the original implementation, if we just add `.jittable()` modify and
call the `torch.jit.script(model)` function, we will get some errors. In
`example/gcn.py`, the error msg is `Expected a proper Tensor but got
None`, and in `example/film.py`, we will get a runtime error: `object
has no attribute or method '__call__'`.

So, we use some small modifications to make these examples really
`jittable`. In the example/film.py, since TorchScript doesn't support
the index loop of `ModuleList`, we split the the last conv layer. In the
`example/gcn.py`, we add a type hint for `edge_weight` variable. And we
kept everything else the same.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>