Add `RemoveDuplicatedEdges` transform (#6709)
Removes duplicate edged to the given homogeneous or heterogeneous graph.
It will change the original order of dataset by concatenating one of
duplicated edges at the end of the dataset. It can be used to clean-up a
known repeated self-connecting edges issue in ogbn-products. Reference
to ogbn-products Leaderboard:
[here](https://ogb.stanford.edu/docs/nodeprop/#:~:text=Note%3A%20A%20very%20small%20number%20of%20self%2Dconnecting%20edges%20are%20repeated%20(see%20here)%3B%20you%20may%20remove%20them%20if%20necessary)

Moved this to separate PR from #6699

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>