Remove adding `num_edges` directly to store in the `Pad` transform (#6758)
Handling `num_edges` as an attribute requires couple more changes and
doesn't work well with the current implementation. For example, when
`num_edges` is set on the `Data` objects and those objects are merged
into a `Batch` the `num_edges` attribute is a tensor with values from
all the objects. It works well when implemented as a property of
`EdgeStorage` and not as a direct attribute, so removing it for now.