Fix `cugraph` tests (#6907)
need more tolerance for cugraph-ops
also for cugraph conversion, we NEED to put the cugraph tensor on cpu or
tests fail. this is recommended and required by the cugraph engineers
```
>       assert torch.equal(edge_index, cu_edge_index)
E       RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cpu and cuda:0! (when checking argument for argument other in method wrapper_CUDA__equal)
```