Let `RandomLinkSplit` work on non-contiguous representations (#6968)
Hi, this PR aims to fix the unexpected crashes on Google Colab. When
using the customized dataset to build the `HeteroData()`, if the edge
lists are not contiguous, an unexpected crash will occur without any
warning messages. Hence, I added a precheck here.

More details can be found in [ISSUE
#6925](https://github.com/pyg-team/pytorch_geometric/issues/6925).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>