feat: add `RotatE` KGE model (#7026)
[No longer contributing affiliated with CS224W]

Paper reported FB15k-237 H@10: .533
Implementation Example H@10: .431

Of note, the paper only reports their score using their Adversarial
Negative Sampling (ANS), which is not implemented as a part of this PR.
Table 7 from the paper comparing the use of ANS vs. uniform random
sampling shows a 5% increase in H@10 for FB15k-237, which may be a
factor to the disparity in performance, combined with the small hidden
size in the example, I believe. The margin=9 argument is from the
paper's reported best performance in Table 12 on FB15k-237 as well.

Happy to try to implement the ANS as part of this PR if the current
performance is below standard, let me know!

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jinu Sunil <jinu.sunil@gmail.com>