`SortAggregation`: Added an optional `max_num_nodes` argument (#7367)
- Added an optional `max_num_nodes` argument to the `SortAggregation`
layer
- Added the possibility to pass the `fill_value` argument as
`torch.tensor` to the `utils.to_dense_batch` function.
- Remove converting `fill_value` to Python float in SortAggregation
forward. Pass `fill_value` as `torch. tensor` to the
`utils.to_dense_batch` function. Reducing the number of Python
operations. Converting to a Python value may cause graph breaks and
splits during torch.compile and as a result the need for the host
instead device calculations (in that case x.min().item() - 1) on some
architectures which which leads to performance degradation.

---------

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>