Add option to override the `segmm` heuristic decision in `HeteroLinear` (#7474)
We added an option to override the heuristic that decides whether segmm
should be used in HeteroLinear
([#7258](https://github.com/pyg-team/pytorch_geometric/pull/7258)).

We also added the option to the constructor of HGTConv which is passed
to the projections for keys and values, respectively.

We found that on one use case we are working on the heuristic
incorrectly selects the for-loop implementation even though it runs
around 3 times slower than the fused segmm kernel.
This happens with a standard 4-layer HGT model with dimension 128 and 4
heads when training on a heterogeneous graph with 16 node types, 26 edge
types and around 1 Million vertices. We can not share the dataset, but
the user should be able to force a certain selection if the heuristic
makes a sub-optimal choice on a given configuration.

Interestingly, we use an A40 GPU which should be very similar to the
A100 that was used to fit the heuristic. It is not clear to us why a
wrong selection is made. We were wondering if the training of the
heuristic should also incorporate the time needed for a backwards pass
through the layer as well as imbalanced node types?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>