Add `Performer` to `GPSConv` (#7465)
### Description

* Most code is adopted and simplified from
https://github.com/lucidrains/performer-pytorch/blob/main/performer_pytorch/performer_pytorch.py
and
https://github.com/rampasek/GraphGPS/blob/main/graphgps/layer/gps_layer.py
Should we add some comments to cite? Checked that both of them have the
MIT License.
* There is a projection matrix redrawer that redraws the projection
matrix in the performer. It seems to be not required so I just put it
into the `examples/graph_gps.py`. Not sure where should I put it.
* There are also some other params and kernels to be added to the
performer, such as softmax kernel etc. For now I added some TODOs.

### ZINC Example
```python
python3 examples/graph_gps.py --attn_type performer
```
I directly replace the multihead attention with performer fast attention
with the `ReLU` kernel. The performance is similar (there are some
randomness for each run and for the graph classification task) Following
is the `test_mae` plot for each epoch.


![training](https://github.com/pyg-team/pytorch_geometric/assets/21955420/50aac085-460c-4293-b74c-966d4857a05e)

---------

Co-authored-by: Jinu Sunil <jinu.sunil@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>