Add `interval` argument to `LocalCartesian` (#7533)
The intermediate tensors `cart` and `max_value` in the original code
were replaced with in-place operations to reduce memory usage. This was
done by directly operating on the `cart` tensor and computing the
maximum value iteratively without creating a separate `max_value`
tensor. In-place operations `(torch.sub, cart.div_, cart.mul_,
cart.add_)` were used to perform computations directly on tensors,
reducing memory usage and eliminating the need for intermediate tensors.
To compute the maximum value in a streaming fashion, a loop was
introduced which iterates over the edges and updates the maximum value
tensor (max_value) accordingly.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>