Update imbalanced_sampler.py
Can we compute the class weights once in the compute_class_weights method and take the dataset and input nodes as arguments? The computed class weights are then stored in the class_weights attribute of the ImbalancedSampler instance.
When creating an instance of ImbalancedSampler, the cached class weights are used to initialize the WeightedRandomSampler superclass, eliminating the need to compute the class weights every time.