[Documentation] make documentation of `GATConv` and `GATv2Conv` more explicit and accurate (#8201)
## What?
I explicitly named the different `\Theta`s used in the different layers
by an index s and t (like it is already done in the docs when talking
about input sizes). I did the same for a in the GATconv layer. Finally
removed the unclear `||` operator and just wrote out the few additions.

## Why?
To solve the issues discussed in
https://github.com/pyg-team/pytorch_geometric/issues/8200

## Testing?
No tests needed, as only documentation was changed.

---------

Co-authored-by: Jintang Li <cnljt@outlook.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>