Allow for device agnostic testing (#8223)
This PR changes the test suite to allow running the tests on any
hardware.

The change is rather small and targets the tests that run both on the
cpu and gpu. For these tests the decorator is changed to also read from
a couple of environment variables that can specify the device and
optionally the backend.
Using environment variables has the benefit of avoiding further changes
to the code while also providing some degree of future proofing. If in
the future new hardware vendors want to run the PyG test suite on their
hardware they simply can use the environment variables.

Usage:
`TORCH_DEVICE=some_new_hw TORCH_BACKEND=torch_new_hw pytest test`

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>