Enable edge based temporal sampling in `torch_geometric.distributed` (#8428)
This PR enables edge based temporal distributed training for node and
link sampling.

**Comment about the edge temporal data definition:**
In the case of distributed training, it is necessary to create a
separate vector for each partition that will store the time information
of the edges included in the partition. (I mention this just to point
out that this works differently than with node-based temporal sampling,
where we can have one vector common to each partition because we operate
on node ids.)
Why:
Each partition has its own unique edge_index in COO format, which is
later converted to a matrix in CSR/CSC format in the neighbor sampler.
Therefore, we do not have information about the global edge IDs when
sampling and we would not be able to find the correct time information
for a specific edge. Therefore, this information must be local.

**Changes made:**
- added support for edge_time argument
- `seed_time` needs to be specified (requirement for edge level temporal
sampling)
- added unit tests
- unit tests for link sampler are in [this PR
#8375](https://github.com/pyg-team/pytorch_geometric/pull/8375)

---------

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>