`NDCG@k` metric for link-prediction (#8326)
Hello!

This branch implements an example of link prediction using the
[ndcg@k](https://en.wikipedia.org/wiki/Discounted_cumulative_gain)
metric.

This has been discussed in the #8271 issue. 

I have tried to first implement it in a new example as a starting point.
It probably should be moved somewhere else, or have a more distinctive
example. This example is a simple copy from the `link_prediction.py`
example, replacing AUC with NDCG.

There remain some questions about the signature of the `update` method,
as well as how to call the method in the example.

For example, we might want to demonstrate that we can call it during
training, adding mini-batches one at a time.

We might also want to avoid predicting links from another set, that way,
we are not considering valid links as undesired.

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>