Enable homo distributed link sampling (#8375)
This implementation covers distributed homo edge sampling. Hetero is in
progress.

The edge_sample() function is almost the same as the one from the
`NeighborSampler` class. The difference is that it calls the
node_sample() from the distributed package.
Why I decided to copy paste it to the DistSampler?:
- to not interfere with the original edge_sample() function from the
`NeighborSampler`
- To make this function async (seamlessly)
- In case of hetero there will be more changes needed

This PR introduces support for edge sampling:
- disjoint, negative, temporal, bidirectional
- with unit tests 
- unit tests for edge-level temporal sampling

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>