chore(pre-commit): migrate to pre-commit for line endings (#8801)
This PR aims to migrate to pre-commit to enforce consistent line
endings. The previous approach introduced in
https://github.com/pyg-team/pytorch_geometric/pull/7759 relies on using
a `.gitattributes` file. Instead this PR proposes to use the
`mixed-line-ending` hook to do the same.

This change was tested by running `pre-commit run --all-files`.

Request for Review: @rusty1s 

I'm skipping editing the `CHANGELOG` for now.