feat(ci/labeler): bump `actions/labeler` and add `skip-changelog` label (#8802)
This PR aims to make the CI process better for contributors by adding
the ability to auto-assign the `skip-changelog` label by naming the
branch `skip*`. With the [new version of `actions/labeler` we can
specify a list of regex
expressions](https://github.com/actions/labeler?tab=readme-ov-file#match-object)
to match with for the head branch.

This will prevent failing CI in case where updating the `CHANGELOG`
isn't considered necessary.

Perhaps this should also be reflected in the docs somewhere, most likely
the contributing guide. More than happy to update the docs as well.

Request for Review: @rusty1s

---------

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>