Rectify the default attention mode "node" of the `nn.HypergraphConv` layer (#8818)
Greetings, esteemed individuals.

While acquainting myself with the utilization of the nn.HypergraphConv
layer through the application of mock inputs, I have encountered an
issue with the default attention mode, namely "node," which disrupts the
functionality of the forward function.

Upon examining the source code, I have identified a potential problem
with the num_nodes parameter within the softmax function at line 177.
From my perspective, when attempting to compute attention scores for
nodes within a hyperedge, the number of hyperedges should be considered
as num_nodes.

Therefore, I have modified the value of num_nodes to correspond to the
pre-calculated quantity of num_edges. As a result, the issue has been
resolved, and the functionality has been restored harmoniously.

Warmest regards.