Fix CI error with socket bind (#8981)
Fixed a repetitive issue in distributed tests with a process binding to
a socket that is already in use:
`torch.distributed.DistNetworkError: The server socket has failed to
listen on any local network address. The server socket has failed to
bind to [::]:54902 (errno: 98 - Address already in use). The server
socket has failed to bind to 0.0.0.0:54902 (errno: 98 - Address already
in use).`

by changing the `localhost` to an empty address which allows any local
address to access the TCP port and adding socket opt 'SO_REUSEADDR':
```
    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
        s.settimeout(1)
        s.bind(('', 0))
        port = s.getsockname()[1]
```

As a bonus I also added a minor change in `test_model_hub.py` that was
causing CI to fail.

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>