refactor(testing): replace withCUDA decorator with withDevice
This commit replaces the usage of the `withCUDA` decorator with the `withDevice` decorator across multiple test files. The `withDevice` decorator provides more flexibility as it allows tests to be run on different processors (cpu only, cpu + gpu, unified memory), enhancing the testing capabilities of the module by allowing tests to be run on different devices.