Added the `residual` option for `GATConv` and `GATv2Conv`. (#9515)
As discussed here
[https://github.com/pyg-team/pytorch_geometric/discussions/7555#discussioncomment-9973267](https://github.com/pyg-team/pytorch_geometric/discussions/7555#discussioncomment-9973267),
the GAT paper mentioned using skip-connection to reach the reported
metrics on PPI dataset. This PR adds the option of `residual` to enable
skip-connection.

Specifically, we consider the following for adding residual:
1. Define the `actual_out_channels` = `heads` * `out_channels` if concat
else `out_channels`
2. If `in_channels` == `actual_out_channels`, then residual is just `x`
before the convolution. Otherwise, residual is a linear projection of
the input `x` onto the dimension of `actual_out_channels`.

For bipartite case where the input is (x_src, x_dst), the above
consideration will be applied for x_dst.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>