Add OPF solution objective function value as a graph level property (#9569)
As discussed with @mzgubic 

This PR has to do with the recently added OPFDataset. It reads the OPF
solution objective function value from the json file, and adds it as a
graph level property. This value is useful in creating optimality
metrics that illustrate how far an ML predictions objective function is
from the one achieved by a solver.