Support ONNX export for `scatter` with min/max reduction (#9587)
Attempt to address issue
https://github.com/pyg-team/pytorch_geometric/issues/8415 and unblock
ONNX export.

As I understand it, the issue is that when we're applying min/max
reductions with `scatter_reduce_` to populate a `self` tensor of zeroes
or an empty tensor, we need to set `include_self=False` to prevent the
initial fill value from influencing the min/max results. For that reason
simply setting `include_self=True` in these cases is not the right
solution since it would change the behaviour of the reduction.

A simple behaviour-preserving solution is to initialize the `self`
tensor such that it doesn't influence the min/max result when
`include_self=True`. This can be done using the dtype min/max values
provided by `torch.finfo` or `torch.iinfo` to initialize the elements of
`self` that will be populated by `scatter_reduce_`. Unfortunately they
aren't currently scriptable (see
https://github.com/pytorch/pytorch/issues/25661), but it's possible to
get them in a roundabout way by casting `torch.inf` (which also works
for int types).

There was a nuance where certain test cases expected zeros in the `self`
tensor for entries that were not involved in the scatter-reduce index.
We can preserve that behaviour by scattering the dtype min/max value
into a zero tensor along the index so that only the relevant entries are
initialized that way.

Another less efficient option would be to calculate the max/min value of
the `src` tensor itself and use that as the fill value.

---------

Co-authored-by: Lee Clement <lee.clement@oxa.tech>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>