Delete test/datasets/test_web_qsp_dataset.py
this test takes too long to run and is only run with FULLTEST which the normal pyg ci doesnt run. in our own NVIDIA CI we test the full g-retriever example which does the same thing as this test since this test only defines the dataset and ensures it was made properly, which our CI running the example ensures.