Sort of working graph generation
Graphs on benchmark dataset are able to somewhat reconstruct original graph structure. However, this only works on sufficiently low thresholds (ie have to set the threshold scale scalar to < 1.5 when the paper suggests > 2). Will have to read the paper further, as I suspect there’s still something off about the thresholding.

See debugging comments to recreate near-perfect original graphs

Results on Mnist dataset still look quite different even when rendered with the debugging comment options, but the average degree of the output graph is still quite similar to the original, so we could be close. This is most likely due to the fact that the thresholding calculation only selects 1-2 singular values, as opposed to over 20 for the benchmark dataset.