Started adding REDDIT-BINARY to test the gmixup at the end of the notebook. Not quite done with that.
I am not sure whether the thresholding is right or wrong. Look at the heatmaps of graphons compared with the example matrices:
they look off...

Maybe there is an issue in the `num_edges` calculation -- should it be divided by two?

Added option to only approximate graphons from a random sample of graphs in the class instead of all of the graphs.