Added gmixup_blog.ipynb which compares two models, one with gmixup and one without. Currently, the model that was trained with GMixup fails, performing much worse than the baseline model... Inspecting the synthetic graphs, they have much lower degrees than real graphs, and much fewer edges, appearing out of distribution, so there might be something wrong with the graphons.
Removed the batching code for graph_generation (sorry for the bad guidance) because it was not clean for the user experience.
The synthetic graphs needed to be shaped and changed after they are generated (add labels, etc) which is easier to do when you generate them one batch at a time.
This feels like the more natural use case and I don't know how much extra utility the batching idea would add for how much it complicated the code.