feat(transforms): add `mesh_to_dual_graph` transformation
This commit introduces a new transformation that converts a mesh to its
"dual".

The dual of a mesh was used implicitly in
[MeshCNN](https://arxiv.org/abs/1809.05910), and referenced in
[PD-MeshNet](https://arxiv.org/abs/2010.12455).

An implementation for this transform can be found in
[pymesh](https://pymesh.readthedocs.io/en/latest/api_misc.html?highlight=dual#pymesh.mesh_to_dual_graph).

The data for this simple test, including the output dual tensor we
validate against, was constructed using Blender, which also provides a
mesh_to_dual transform (can be found as a geometry node).