Adds checkpointing for KG creation to `tech_qa.py` (#10135)
Based on code from
[here](https://github.com/pyg-team/pytorch_geometric/blob/bb3992d7d488af576c948c267ebfe79b97227496/examples/llm/hotpot_qa.py#L32),
this adds a `--checkpointing` option to the `tech_qa.py` script which
saves the progress of KG creation at a set interval, allowing for the
restoration of the KG in order to resume from a point close to where the
script quit unexpectedly.

This has only been "lightly" tested using `pdb` to step through and
inspect state. This still needs a proper end-to-end run with an
interruption to verify the KG is properly restored and creation resumed
from a checkpoint.

Co-authored-by: Rishi Puri <riship@nvidia.com>