GPSE example (#10118)
Following #9018, Provides a comprehensive example in `examples/gpse.py`
to compute GPSE encodings and use them for a graph regression task on
the ZINC dataset. Two methods to compute GPSE encodings is demonstrated:

- Through `precompute_GPSE`: Given a PyG dataset, computes GPSE
encodings in-place _once_ before training, without saving them to
storage. Ideal if you want to compute the encodings only once per run
(unlike a dataset transform) but do not want to save the pre-transformed
dataset to storage (unlike a PyG pre-transform).

To run with default pretrained weights (molpcba):
```
python examples/gpse.py --gpse
```
To run with pretrained weights from any other dataset, please provide
the pretraining dataset name from the available options as a kwarg:
```
python examples/gpse.py --gpse geom
```

- Through the `AddGPSE` transform: A PyG transform analogous to
[AddLaplacianEigenvectorPE](https://pytorch-geometric.readthedocs.io/en/2.6.0/generated/torch_geometric.transforms.AddLaplacianEigenvectorPE.html#torch_geometric.transforms.AddLaplacianEigenvectorPE)
and
[AddRandomWalkPE](https://pytorch-geometric.readthedocs.io/en/2.6.0/generated/torch_geometric.transforms.AddRandomWalkPE.html#torch_geometric.transforms.AddRandomWalkPE),
can be used as a pre-transform or transform to a PyG dataset.
```
python examples/gpse.py --gpse --as_transform
```

Using as a transform is not recommended as recomputing them for every
batch in every epoch is quite inefficient; using it as a pre-transform
or through `precompute_GPSE` is suggested instead. In either case, the
`torch_geometric.nn.GPSENodeEncoder` is then used to compute a mapping
of the GPSE encodings to the desired dimension, and append them to
`batch.x` to prepare them as inputs to a GNN.

This PR has been tested with the latest (25.01-py3) NVIDIA stack, and
has works without any issues.

---------

Co-authored-by: Semih Cantürk <=>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Rishi Puri <puririshi98@berkeley.edu>
Co-authored-by: Rishi Puri <riship@nvidia.com>