Bi-directional Sampling for NeighborSampler (#10126)
This PR makes the following changes:

- Modifies NeighborSampler so that sampling can occur from `src` to
`dst` node, and not just from `dst` to `src` node, which is the current
default behavior.
- Adds a new sampler called `BidirectionalNeighborSampler`, which
samples nodes/edges in both upstream and downstream directions for a
directed graph, as if it were an undirected graph.
- Add `merge_with` and `collate` methods to the SamplerOutput dataclass,
allowing for different SamplerOutputs to be merged together. This also
becomes useful in scenarios where the user may want to sample using
multiple different unique sampling methods.

This PR will remain in draft until ~methods are also implemented on the
corresponding Hetero graph classes, and~ unittests have been written to
cover new behaviors.
UPDATE: Hetero Bidirectional Sampler and SamplerOutput merge will be a
separate PR.

---------

Co-authored-by: Zack Aristei <zaristei@zaristei-nvidia.client.nvidia.com>
Co-authored-by: Rishi Puri <riship@nvidia.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>